[rename] 파일 이동으로 인한 경로 수정
feat: 새로운 기능 추가
fix: 버그 수정
docs: 문서 수정
design: 퍼블리싱, CSS 등 사용자 UI 디자인 변경
style: 들여쓰기, 세미콜론 등 코드 스타일 변경
refactor: 코드 리팩토링
chore: 외부 라이브러리 임포트 등의 작업 완료
comment: 필요한 주석 추가 및 변경
rename: 파일 혹은 폴더명을 수정하거나 옮기는 작업만 한 경우
remove: 파일을 삭제하는 작업만 한 경우
